+++ app/routes/blog.$slug.tsx
@@ -5,7 +5,7 @@ import { getMDXComponent } from "mdx-bundler/client";
 import { useMemo } from "react";
 import invariant from "tiny-invariant";
 import { PUBLISH_DATE_FORMATTER } from "~/utils";
-import type { Mdx } from "blog/parser";
+import type { Mdx } from "blog/parser.server";
 
 export const loader = async function ({ params, request }: LoaderArgs) {
   const { slug } = params;
