+++ package.json
@@ -45,8 +45,8 @@
   ],
   "dependencies": {
     "@remix-run/node": "^1.15.0",
-    "@remix-run/react": "^1.15.0",
-    "@remix-run/serve": "^1.15.0",
+    "@remix-run/react": "^1.16.1",
+    "@remix-run/serve": "^1.16.1",
     "daisyui": "2.38.0",
     "isbot": "^3.5.3",
     "mdx-bundler": "^9.2.1",
@@ -58,8 +58,8 @@
   },
   "devDependencies": {
     "@playwright/test": "1.28.1",
-    "@remix-run/dev": "^1.15.0",
-    "@remix-run/eslint-config": "^1.15.0",
+    "@remix-run/dev": "^1.16.1",
+    "@remix-run/eslint-config": "^1.16.1",
     "@tailwindcss/typography": "0.5.7",
     "@testing-library/dom": "^8.18.1",
     "@testing-library/jest-dom": "^5.16.5",
